// src/types/rows/OrcamentoCabecalhoReal.ts
export interface OrcamentoCabecalhoReal {
  anomes: string;
  faturamento_bruto_real: number;
  imposto_vendas_real: number;
  cmv_real: number;
  diferenca_inv_real: number;
  lucro_bruto_real: number;
  despesas_real: number;
  perdas_real: number;
  receita_financ_real: number;
  receita_comercial_real: number;
  receita_marketing_real: number;
  lucro_operacional_real: number;
  irpj_real: number;
  csll_real: number;
  lucro_liquido_real: number;
}